+++ src/Routing/LocaleRouter.php
@@ -4,9 +4,11 @@ namespace CaribouFute\LocaleRoute\Routing;
 
 use CaribouFute\LocaleRoute\Prefix\Route as PrefixRoute;
 use CaribouFute\LocaleRoute\Prefix\Url as PrefixUrl;
+use CaribouFute\LocaleRoute\Routing\ResourceRegistrar;
 use CaribouFute\LocaleRoute\Routing\Router;
 use CaribouFute\LocaleRoute\Traits\ConfigParams;
 use CaribouFute\LocaleRoute\Traits\ConvertToControllerAction;
+use Illuminate\Translation\Translator;
 
 class LocaleRouter
 {
@@ -98,5 +100,12 @@ class LocaleRouter
         }
 
         return $action;
+    }
+
+    public function resource($route, $controller, $options = [])
+    {
+        $registrar = new ResourceRegistrar($this, $this->router->getRouter(), app()->make(Translator::class));
+
+        $registrar->register($route, $controller, $options);
     }
 }